+++ src/lib/application/services/index.ts
@@ -11,3 +11,5 @@ export { GitIntrospectionService } from './GitIntrospectionService';
 
 export { MemoryContextLoader } from './MemoryContextLoader';
 export type { IMemoryLoadResult } from './MemoryContextLoader';
+
+export { GitTriageService } from './GitTriageService';